% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_studies.R
\name{count_studies}
\alias{count_studies}
\title{Count the number of unique studies included in a tibble}
\usage{
count_studies(data, nct_column = "nct_id")
}
\arguments{
\item{data}{A tibble}

\item{nct_column}{The name of the column of \code{data} that corresponds to
the NCT ID. Defaults to \code{"nct_id"}.}
}
\value{
A count of the number of unique NCT IDs in the table, as a double.
}
\description{
Count the number of unique studies included in a tibble
}
