% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctdl-package.R
\docType{package}
\name{ctdl-package}
\alias{ctdl}
\alias{ctdl-package}
\title{ctdl: Clinical Trials Downloader}
\description{
| This package downloads clinicaltrials.gov data from the Clinical Trials Transformation Initiative (CTTI) (https://ctti-clinicaltrials.org/).
}
\author{
\strong{Maintainer}: Brendan O'Leary \email{brendan@boleary.com} (\href{https://orcid.org/0000-0001-8753-917X}{ORCID})

}
\keyword{internal}
