% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download CTTI Data}
\usage{
download_data(filepath_data = "data-raw/")
}
\arguments{
\item{filepath_data}{The location to download the archive to. This defaults
to \code{data-raw/}. /If this path already exists, it and its contents will
be archived and the directory will be created fresh.}
}
\value{
The date associated with the data file downloaded.
}
\description{
Downloads the most recent archive of clinicaltrials.gov from the CTTI site.
}
