% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_studies.R
\name{find_studies}
\alias{find_studies}
\title{Find Studies}
\usage{
find_studies(data_studies, search_terms, known_fp = NULL, max_topics = 5L)
}
\arguments{
\item{data_studies}{A tibble including the columns \code{official_title},
\code{brief_title}, and \code{nct_id}.}

\item{search_terms}{A tibble including the columns \code{regex_term} and
\code{topic}.}

\item{known_fp}{A vector of NCT numbers that should be excluded from the
search results. Defaults to \code{NULL}.}

\item{max_topics}{An integer value for the maximum number of topics to
include in the results. Less frequent topics beyond this number will be
collapsed into an "Other" topic.}
}
\value{
A tibble.
}
\description{
Find Studies
}
