% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_studies.R
\name{plot_studies}
\alias{plot_studies}
\title{Plot Studies}
\usage{
plot_studies(studies, description, this_year, search_terms)
}
\arguments{
\item{studies}{A tibble of studies}

\item{description}{A descriptive term about the data in \code{studies} to be
included in the plot title and captions.}

\item{this_year}{This year.}

\item{search_terms}{The terms used to filter the contents of \code{studies}.}
}
\value{
A plot.
}
\description{
Plot Studies
}
